Views secured and restricted to admins

+++ shop/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Cd
 from .forms import CdForm
 
@@ -25,7 +26,12 @@ def cd_detail(request, cd_id):
     return render(request, "shop/cd_details.html", context)
 
 
+@login_required
 def add_cd(request):
+    if not request.user.is_superuser:
+        messages.error(request, "Space reserved to administrators.")
+        return redirect(reverse("home"))
+
     if request.method == "POST":
         form = CdForm(request.POST, request.FILES)
         if form.is_valid():
@@ -45,7 +51,12 @@ def add_cd(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_cd(request, cd_id):
+    if not request.user.is_superuser:
+        messages.error(request, "Space reserved to administrators.")
+        return redirect(reverse("home"))
+
     cd = get_object_or_404(Cd, pk=cd_id)
     if request.method == "POST":
         form = CdForm(request.POST, request.FILES, instance=cd)
@@ -71,7 +82,12 @@ def edit_cd(request, cd_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_cd(request, cd_id):
+    if not request.user.is_superuser:
+        messages.error(request, "Space reserved to administrators.")
+        return redirect(reverse("home"))
+
     cd = get_object_or_404(Cd, pk=cd_id)
     cd.delete()
     messages.success(request, "CD deleted.")
